rename folder and files
